refactor(cache): split schema responsibilities
